requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.0
Pillow>=9.0.0
GDAL>=3.4.0
tqdm>=4.66.0
PyYAML>=6.0.0
opencv-python>=4.8.0
matplotlib>=3.7.0